refactor(workspace, svelte, apps): collapse userId plumbing into LocalOwner
Introduce createLocalOwner({ userId, encryptionKeys }) in @epicenter/workspace
as the identity-scoped facade for authenticated browser workspaces. The owner
owns the four verbs that depended on the signed-in user's local scope:
attachEncryption(ydoc), attachIndexedDb(ydoc), attachBroadcastChannel(ydoc),
and wipeLocalYjsData(guids).

createSession's build context shrinks from { identity, encryptionKeys } to
{ owner }. Browser bundles receive owner once and forward it; no more
identity.user.id extraction, no more encryptionKeys callback threaded through
five files, no more userId argument on every owner-aware workspace call.

Daemons are untouched: they call the free attachEncryption(ydoc, { encryptionKeys })
directly and persist via filesystem, so the owner concept never enters that
path.

Deletes the standalone attachOwnedBroadcastChannel and wipeOwnerLocalYjsData
exports. Encrypted IndexedDB roundtrip tests move from attach-encryption.test.ts
to local-owner.test.ts, where they belong.
