+++ apps/honeycrisp/src/routes/(signed-in)/honeycrisp/browser.ts
@@ -5,18 +5,15 @@ import {
 	type NoteId,
 } from '@epicenter/honeycrisp';
 import {
-	attachEncryption,
-	attachOwnedBroadcastChannel,
 	attachRichText,
 	createDisposableCache,
 	DateTimeString,
 	docGuid,
-	type EncryptionKeys,
+	type LocalOwner,
 	type OpenWebSocket,
 	onLocalUpdate,
 	openCollaboration,
 	websocketUrl,
-	wipeOwnerLocalYjsData,
 } from '@epicenter/workspace';
 import * as Y from 'yjs';
 
@@ -36,23 +33,21 @@ function noteBodyDocGuid({
 }
 
 export function openHoneycrispBrowser({
-	userId,
+	owner,
 	replicaId,
 	openWebSocket,
-	encryptionKeys,
 }: {
-	userId: string;
+	owner: LocalOwner;
 	replicaId: string;
 	openWebSocket?: OpenWebSocket;
-	encryptionKeys: () => EncryptionKeys;
 }) {
 	const rootYdoc = new Y.Doc({ guid: 'epicenter.honeycrisp', gc: false });
-	const encryption = attachEncryption(rootYdoc, { encryptionKeys });
+	const encryption = owner.attachEncryption(rootYdoc);
 	const tables = encryption.attachTables(honeycrispTables);
 	const kv = encryption.attachKv({});
 
-	const idb = encryption.attachIndexedDb(rootYdoc, { userId });
-	attachOwnedBroadcastChannel(rootYdoc, { userId });
+	const idb = owner.attachIndexedDb(rootYdoc);
+	owner.attachBroadcastChannel(rootYdoc);
 
 	const noteBodyDocs = createDisposableCache((noteId: NoteId) => {
 		const ydoc = new Y.Doc({
@@ -63,8 +58,8 @@ export function openHoneycrispBrowser({
 			gc: false,
 		});
 		const body = attachRichText(ydoc);
-		const childIdb = encryption.attachIndexedDb(ydoc, { userId });
-		attachOwnedBroadcastChannel(ydoc, { userId });
+		const childIdb = owner.attachIndexedDb(ydoc);
+		owner.attachBroadcastChannel(ydoc);
 		const childSync = openCollaboration(ydoc, {
 			url: websocketUrl(`${APP_URLS.API}/documents/${ydoc.guid}`),
 			waitFor: childIdb.whenLoaded,
@@ -124,10 +119,7 @@ export function openHoneycrispBrowser({
 			noteBodyDocs[Symbol.dispose]();
 			rootYdoc.destroy();
 			await Promise.all([idb.whenDisposed, collaboration.whenDisposed]);
-			await wipeOwnerLocalYjsData({
-				userId,
-				ydocGuids: fallbackGuids,
-			});
+			await owner.wipeLocalYjsData(fallbackGuids);
 		},
 		[Symbol.dispose]() {
 			noteBodyDocs[Symbol.dispose]();
