+++ apps/opensidian/src/lib/opensidian/browser.ts
@@ -6,16 +6,13 @@ import {
 	fileContentDocGuid,
 } from '@epicenter/filesystem';
 import {
-	attachEncryption,
-	attachOwnedBroadcastChannel,
 	attachTimeline,
 	createDisposableCache,
-	type EncryptionKeys,
+	type LocalOwner,
 	type OpenWebSocket,
 	onLocalUpdate,
 	openCollaboration,
 	websocketUrl,
-	wipeOwnerLocalYjsData,
 } from '@epicenter/workspace';
 import { Bash } from 'just-bash';
 import { opensidianTables } from 'opensidian';
@@ -23,23 +20,21 @@ import * as Y from 'yjs';
 import { createOpensidianActions } from './actions';
 
 export function openOpensidianBrowser({
-	userId,
+	owner,
 	replicaId,
 	openWebSocket,
-	encryptionKeys,
 }: {
-	userId: string;
+	owner: LocalOwner;
 	replicaId: string;
 	openWebSocket?: OpenWebSocket;
-	encryptionKeys: () => EncryptionKeys;
 }) {
 	const rootYdoc = new Y.Doc({ guid: 'epicenter.opensidian', gc: false });
-	const encryption = attachEncryption(rootYdoc, { encryptionKeys });
+	const encryption = owner.attachEncryption(rootYdoc);
 	const tables = encryption.attachTables(opensidianTables);
 	const kv = encryption.attachKv({});
 
-	const idb = encryption.attachIndexedDb(rootYdoc, { userId });
-	attachOwnedBroadcastChannel(rootYdoc, { userId });
+	const idb = owner.attachIndexedDb(rootYdoc);
+	owner.attachBroadcastChannel(rootYdoc);
 
 	const fileContentDocs = createDisposableCache((fileId: FileId) => {
 		const ydoc = new Y.Doc({
@@ -52,8 +47,8 @@ export function openOpensidianBrowser({
 		onLocalUpdate(ydoc, () =>
 			tables.files.update(fileId, { updatedAt: Date.now() }),
 		);
-		const childIdb = encryption.attachIndexedDb(ydoc, { userId });
-		attachOwnedBroadcastChannel(ydoc, { userId });
+		const childIdb = owner.attachIndexedDb(ydoc);
+		owner.attachBroadcastChannel(ydoc);
 		return {
 			ydoc,
 			content: attachTimeline(ydoc),
@@ -140,10 +135,7 @@ export function openOpensidianBrowser({
 			];
 			disposeResources();
 			await Promise.all([idb.whenDisposed, collaboration.whenDisposed]);
-			await wipeOwnerLocalYjsData({
-				userId,
-				ydocGuids: fallbackGuids,
-			});
+			await owner.wipeLocalYjsData(fallbackGuids);
 		},
 		[Symbol.dispose]() {
 			disposeResources();
