+++ packages/workspace/src/document/attach-encryption.ts
@@ -24,63 +24,61 @@
  *
  * ## Key source: lazy callback
  *
- * `encryptionKeys` is a callback into whoever owns identity.
- * The coordinator calls it synchronously at every `attachTable` / `attachKv` /
- * `attachIndexedDb` site, derives the keyring, and activates the store. The
- * keyring is not cached on the attachment: each attach call is its own
- * derivation, which keeps state out of this layer entirely.
+ * `encryptionKeys` is a callback into whoever owns identity. The coordinator
+ * calls it synchronously at every `attachTable` / `attachKv` site, derives the
+ * keyring, and activates the store. The keyring is not cached on the
+ * attachment: each attach call is its own derivation, which keeps state out of
+ * this layer entirely.
  *
  * Same-user identity updates (key rotation, profile edits) do not flow
- * through this attachment. The session lifecycle reloads the page on
- * different-user transitions; same-user updates are observed lazily via
- * the `encryptionKeys` callback the next time it runs.
+ * through this attachment. Authenticated apps reload the page on
+ * different-user transitions; same-user updates are observed lazily via the
+ * `encryptionKeys` callback the next time it runs.
+ *
+ * ## Local owner concerns live on `createLocalOwner`
+ *
+ * Encrypted IndexedDB, owner-scoped BroadcastChannel, and local-data wipe are
+ * identity-scoped (one signed-in user) rather than ydoc-scoped, so they live
+ * on `createLocalOwner` (`./local-owner.ts`). Daemons that only need
+ * encryption call `attachEncryption` directly; browsers go through
+ * `LocalOwner`, which delegates encryption to this function.
  *
  * ## Disposal
  *
  * Each attached store hooks `ydoc.once('destroy', ...)` at attach time,
- * mirroring the plaintext `attachTable` / `attachKv` primitives.
- * `attachIndexedDb` is the exception: its IDB provider wires its own
- * teardown inside `attachEncryptedIndexedDb` rather than routing through
- * the shared `attachStore` helper. Callers tear down encryption by calling
- * `ydoc.destroy()`: the attachment does not expose a standalone `dispose()`
- * method.
+ * mirroring the plaintext `attachTable` / `attachKv` primitives. Callers tear
+ * down encryption by calling `ydoc.destroy()`: the attachment does not expose
+ * a standalone `dispose()` method.
  *
  * ## What this attachment does NOT do
  *
- * - It does not wipe CRDT state. Any future "wipe encrypted blobs" API needs
- *   to coordinate with persistence to be useful: design it alongside the
- *   consumer migration.
- * - It does not validate that every encryption-capable slot on the Y.Doc
- *   got registered. The caller owns the composition: if you pair a
- *   plaintext `attachTable` with `encryption.attachTable` targeting the
- *   *same slot name*, Yjs hands both calls the same underlying `Y.Array` and
- *   you get a silent plaintext-over-ciphertext race. The verb
- *   (`encryption.attachTable` vs plain `attachTable`) is the primary defense;
- *   review call sites accordingly. One slot name, one attach site, one intent.
+ * - It does not wipe CRDT state. `LocalOwner.wipeLocalYjsData` owns that.
+ * - It does not validate that every encryption-capable slot on the Y.Doc got
+ *   registered. The caller owns the composition: if you pair a plaintext
+ *   `attachTable` with `encryption.attachTable` targeting the *same slot
+ *   name*, Yjs hands both calls the same underlying `Y.Array` and you get a
+ *   silent plaintext-over-ciphertext race. The verb (`encryption.attachTable`
+ *   vs plain `attachTable`) is the primary defense; review call sites
+ *   accordingly. One slot name, one attach site, one intent.
  *
  * ## Why `workspaceId` is read from `ydoc.guid`
  *
  * By construction, the workspace Y.Doc's `guid` equals the workspace id
  * (`new Y.Doc({ guid: id })`). Taking a separate `workspaceId` parameter
- * would invite drift between the two. `deriveWorkspaceKey` uses the id as
- * an HKDF domain-separation label: it doesn't care whether the string is
- * the guid or an explicit id, only that the two agree.
+ * would invite drift between the two. `deriveWorkspaceKey` uses the id as an
+ * HKDF domain-separation label: it doesn't care whether the string is the
+ * guid or an explicit id, only that the two agree.
  *
  * @module
  */
 
-import {
-	base64ToBytes,
-	deriveWorkspaceKey,
-	type EncryptionKeys,
-} from '@epicenter/encryption';
+import type { EncryptionKeys } from '@epicenter/encryption';
 import type * as Y from 'yjs';
 import {
 	createEncryptedYkvLww,
 	type EncryptedYKeyValueLww,
 } from '../shared/y-keyvalue/y-keyvalue-lww-encrypted.js';
-import { attachEncryptedIndexedDb } from './attach-encrypted-indexed-db.js';
-import type { IndexedDbAttachment } from './attach-indexed-db.js';
+import { deriveWorkspaceKeyring } from './derive-workspace-keyring.js';
 import { createKv, type Kv, type KvDefinitions } from './attach-kv.js';
 import {
 	createReadonlyTable,
@@ -94,16 +92,14 @@ import {
 	type Tables,
 } from './attach-table.js';
 import { KV_KEY, TableKey } from './keys.js';
-import { createOwnedYjsKey } from './local-yjs-key.js';
 
 export type AttachEncryptionOptions = {
 	/**
 	 * Lazy reader for the current user's encryption keys.
 	 *
-	 * Called synchronously at every `attachTable` / `attachKv` /
-	 * `attachIndexedDb` site. Throw if no keys are available (e.g. signed-out):
-	 * a throw here means the workspace outlived its signed-in scope, which is a
-	 * caller bug.
+	 * Called synchronously at every `attachTable` / `attachKv` site. Throw if
+	 * no keys are available (e.g. signed-out): a throw here means the workspace
+	 * outlived its signed-in scope, which is a caller bug.
 	 */
 	encryptionKeys: () => EncryptionKeys;
 };
@@ -144,27 +140,15 @@ export type EncryptionAttachment = {
 	 * Attach the encrypted KV singleton to the coordinator's Y.Doc.
 	 */
 	attachKv<T extends KvDefinitions>(definitions: T): Kv<T>;
-
-	/**
-	 * Attach encrypted local IndexedDB persistence for a root or child Y.Doc.
-	 *
-	 * Reads keys via `options.encryptionKeys()` at attach time and binds the derived
-	 * keyring to the provider. Same-user key rotation is not observed by the
-	 * provider after this point; cross-user transitions reload the page.
-	 */
-	attachIndexedDb(
-		targetYdoc: Y.Doc,
-		opts: { userId: string },
-	): IndexedDbAttachment;
 };
 
 /**
  * Create an encryption coordinator bound to `ydoc`.
  *
- * The returned coordinator owns `attachTable` / `attachTables` / `attachKv` /
- * `attachIndexedDb` methods: call them to register encrypted resources. The
- * coordinator reads `options.encryptionKeys()` synchronously at each registration
- * site and activates the resource before returning it.
+ * The returned coordinator owns `attachTable` / `attachTables` / `attachKv`
+ * methods: call them to register encrypted resources. The coordinator reads
+ * `options.encryptionKeys()` synchronously at each registration site and
+ * activates the resource before returning it.
  */
 export function attachEncryption(
 	ydoc: Y.Doc,
@@ -172,24 +156,12 @@ export function attachEncryption(
 ): EncryptionAttachment {
 	const workspaceId = ydoc.guid;
 
-	function deriveKeyring(
-		keys: EncryptionKeys,
-		targetWorkspaceId: string,
-	): Map<number, Uint8Array> {
-		const keyring = new Map<number, Uint8Array>();
-		for (const { version, userKeyBase64 } of keys) {
-			const userKey = base64ToBytes(userKeyBase64);
-			keyring.set(version, deriveWorkspaceKey(userKey, targetWorkspaceId));
-		}
-		return keyring;
-	}
-
 	// biome-ignore lint/suspicious/noExplicitAny: variance
 	function attachStore(key: string): EncryptedYKeyValueLww<any> {
 		const store = createEncryptedYkvLww(ydoc, key);
 		ydoc.once('destroy', () => store[Symbol.dispose]());
 		store.activateEncryption(
-			deriveKeyring(options.encryptionKeys(), workspaceId),
+			deriveWorkspaceKeyring(options.encryptionKeys(), workspaceId),
 		);
 		return store;
 	}
@@ -219,21 +191,6 @@ export function attachEncryption(
 		},
 		attachKv(definitions) {
 			return createKv(attachStore(KV_KEY), definitions);
-		},
-		attachIndexedDb(targetYdoc, { userId }) {
-			const keyring = deriveKeyring(options.encryptionKeys(), targetYdoc.guid);
-			if (keyring.size === 0) {
-				throw new Error(
-					'Cannot attach encrypted IndexedDB provider: encryptionKeys() returned no usable keys.',
-				);
-			}
-			const version = Math.max(...keyring.keys());
-			const bytes = keyring.get(version)!;
-			return attachEncryptedIndexedDb(targetYdoc, {
-				databaseName: createOwnedYjsKey(userId, targetYdoc.guid),
-				writeKey: { version, bytes },
-				keyring,
-			});
 		},
 	};
